pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
  "poethepoet>=0.32.0",
  "pre-commit>=4.2.0",
  "pyright>=1.1.391",
  "pytest>=8.3.4",
  "ruff>=0.8.5"
]

[project]
name = "dicom2img"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
  {name = "Jakob Guldberg Aaes", email = "jakob1397@gmail.com"}
]
# Public libraries should be more lenient
# Internal stuff should enforce ~=3.13!
requires-python = "~=3.12"
dependencies = [
  "numpy>=2.2.4",
  "pillow>=11.1.0",
  "pydicom>=3.0.1",
  "rich>=14.0.0",
  "typer>=0.15.2"
]

[project.scripts]
dicom2img = "dicom2img.cli:app"

[tool.poe.tasks]
# run with eg `uv run poe fmt`
fmt = "ruff format"
lint = "ruff check --fix"
check = "pyright"
test = "pytest"
# run all the above
all = [{ref = "fmt"}, {ref = "lint"}, {ref = "check"}, {ref = "test"}]
"ci:fmt" = "ruff format --check" # fail if not formatted
"ci:lint" = "ruff check"

[tool.pyright]
venvPath = "." # uv installs the venv in the current dir
venv = ".venv" # in a folder called `.venv`
strict = ["**/*.py"] # use 'strict' checking on all files
pythonVersion = "3.13" # if library, specify the _lowest_ you support

[tool.pytest.ini_options]
addopts = "--doctest-modules"

[tool.ruff]
# if this is a library, enter the _minimum_ version you
# want to support, otherwise do py313
target-version = "py312"
line-length = 120 # use whatever number makes you happy

[tool.ruff.lint]
# you can see the looong list of rules here:
# https://docs.astral.sh/ruff/rules/
# here's a couple to start with
select = [
  "A", # warn about shadowing built-ins
  "E", # style stuff, whitespaces
  "F", # important pyflakes lints
  "I", # import sorting
  "N", # naming
  "T100" # breakpoints (probably don't want these in prod!)
]

[tool.ruff.lint.isort]
# so it knows to group first-party stuff last
known-first-party = ["dicom2img"]
